Berlin2016-Demo/agile_api_protocol/requirements.txt
#
#######  Pip module requirements  #######
####### for the AGILE DBus Server #######
#
pyserial
RPi.GPIO
dbus-python
XBee

